+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+#### modify the first argument to `#search()` to be either a string or an array
+#### don't add a $text property to the query if it is a blank string or blank array
+
 ## 1.0.0
 
 - Refactor out `_textIndex` properties in favour of using mongo's full text search